Add files via upload
The proposed online appointment system has been implemented in Visual studio and Xampp server for web portal development and it is developed using HTML and PHP. The tasks involved in this work are divided into three modules; Doctor, Patient and Admin. The proposed system is efficient and has a friendly user interface.